separate PSF and noise wrappers for GalSim into two separate files
