+++ python/lsst/sims/catUtils/galSimInterface/__init__.py
@@ -1,3 +1,4 @@
-from .galSimUtilities import *
+from .galSimNoiseAndBackground import *
+from .galSimPSF import *
 from .galSimInterpreter import *
 from .galSimCatalogs import *
