+++ python/lsst/sims/catUtils/galSimInterface/galSimPSF.py
@@ -1,5 +1,5 @@
 """
-This file defines the model classes that wrap PSFs and noise models from
+This file defines the model classes that wrap PSFs from
 galsim into the CatSim interface
 """
 
@@ -7,7 +7,7 @@ import numpy
 import galsim
 from lsst.sims.utils import radiansToArcsec
 
-__all__ = ["PSFbase", "DoubleGaussianPSF", "SNRdocumentPSF", "ExampleCCDNoise"]
+__all__ = ["PSFbase", "DoubleGaussianPSF", "SNRdocumentPSF",]
 
 class PSFbase(object):
     """
@@ -176,38 +176,3 @@ class SNRdocumentPSF(DoubleGaussianPSF):
         gaussian2 = galsim.Gaussian(sigma=2.0*alpha)
 
         self._cached_psf = 0.909*(gaussian1 + 0.1*gaussian2)
-
-
-
-class ExampleCCDNoise(object):
-    """
-    This class wraps the GalSim class CCDNoise.  It is meant to be assigned as
-    the self.noise member variable in a GalSim InstanceCatalog.  To instantiate
-    a different noise model, write a class like this one that defines a method
-    getNoiseModel() which accepts as its argument an ObservationMetaData
-    instantiation (see
-
-    sims_catalogs_generation/python/lsst/sims/catalogs/generation/db/ObservationMetaData.py
-
-    for definition) and a GalSimDetector instantiation and returns an instantiation of
-    a GalSim noise model
-    """
-
-    def __init__(self, seed=None):
-        if seed is None:
-            self.randomNumbers = galsim.UniformDeviate()
-        else:
-            self.randomNumbers = galsim.UniformDeviate(seed)
-
-    def getNoiseModel(self, obs_metadata=None, detector=None):
-        skyLevel = 10.0 #this is obviously nonsense; GalSim wants electrons-per-pixel; Peter thinks we are storing
-                        #sky brightness in magnitudes per square arc-second; once we have the new interface to
-                        #OpSim written, we will need to use that, plus filter information, to convert
-                        #between the two (which, in principle, can be done)
-
-        gain = detector.electronsPerADU
-        readNoise = detector.readNoise
-
-        return galsim.CCDNoise(self.randomNumbers, sky_level=skyLevel, gain=gain, read_noise=readNoise)
-
-
